FileArchiver: Make the storage path independent of any package Id
The content of any file archive only depends on (a subset of the
attributes of) the provenance, but is independent of the actual package
context or rather identifier.  Note that the content of any file archive
is idependent of the VCS path, because the sparse checkout does adhere
to the configured license filename patterns.

Make the file archive paths independent of the identifier to avoid
storing duplicate archives and to move towards using an appropritate
key.

Signed-off-by: Frank Viernau <frank.viernau@here.com>