LicenseInfoResolverTest: Make test resilient to file archiver changes
Previously, the asset for this test contained the directory / file
structure created by the file archiver. Thus, the way how the file
archiver creates its file structure cannot be changed without breaking
this test.

Adjust the test asset so that it contains the actual file to be archived
and create the actual file structure via FileArchiver.archive() to fix
that. As nice side effect the zip file gets eliminated from the test
assests.

Signed-off-by: Frank Viernau <frank.viernau@here.com>